Add support for nested layouts
Adds the "inherits" configuration option to layout.
This allows the user to configure layouts that re-use
the commands in a nested fashion.

This is useful when you want similar layouts but don't want
to re-write the same annoying tmux commands all over the place.

This should be a good ergonomic improvement to the layout system,
making it slightly more in line with expectations one might have
moving over from some of the alternatives to twm.