Fix bug where window resizing breaks if a session is launched via twm
Addresses the issue with gh-16 where resizing the window breaks
everything if your tmux session was attached to via execvp. I still have
*no* idea why this is happening, but this seems to fix it at least;
instead of calling execvp with tmux as the target program, we call it to
whatever the user's shell is with `-c tmux attach -t <session_name>` as
the arguments. Tested this across multiple shells and it seems to work
fine.