Fix double-keystroke bug in layout selection
Currently there is a bug when selecting a layout where you need to press
keys 2-3 times for them to register. I have no idea why, but this fixes
it.

Something about starting a Tui instance to select the workspace, closing
it, and then starting another one to select the layout doesn't work.

This change starts a Tui instance as needed earlier in the program and
passes it around until we're done using it.

Also slightly refactors a couple places.